requirements.txt
aiohttp>=3.9
tqdm>=4.66
Pillow>=10.0
